export/simple_research/requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
uvloop>=0.17.0
